+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 * Define all `Object` pattern instances for `Constant` datatype
 * Define `reset` and `shift` methods for `Continuation` datatype
 * Define `Endo` datatype in `Basis` module
+* Define `Object` instances for transformer schemes